+++ dotnet/NormantonNexus/Models/Dto/ConsignmentTrackerModels.cs
@@ -7,11 +7,11 @@ namespace NormantonNexus.Models.Dto;
 // stale dbo./"kongsberg database" draft.
 
 public sealed record ConsignmentVendorRow(
-    long VendorId, string VendorName, string? SapVendorNumber, string? Currency,
+    int VendorId, string VendorName, string? SapVendorNumber, string? Currency,
     bool TrackExpiry, int? ExpiryWarningDays, int? ExpiryDays, string DefaultAllocationMethod,
     bool Active, string? Notes, DateTime? UpdatedAtUtc, string? UpdatedByUsername);
 
-public sealed record VendorMaterialRow(long VendorMaterialId, string Material, string? ScheduleAgreement);
+public sealed record VendorMaterialRow(int VendorMaterialId, string Material, string? ScheduleAgreement);
 
 public sealed record ConsignmentVendorDetail(ConsignmentVendorRow Vendor, IReadOnlyList<VendorMaterialRow> Materials);
 
@@ -20,7 +20,7 @@ public sealed record UpsertConsignmentVendorConfigRequest(
     bool? TrackExpiry, int? ExpiryWarningDays, int? ExpiryDays, string? DefaultAllocationMethod, bool? Active, string? Notes);
 
 public sealed record ConsignmentDeliveryRow(
-    long DeliveryId, long VendorId, string Material, string MaterialDocument, string MaterialDocItem,
+    int DeliveryId, int VendorId, string Material, string MaterialDocument, string MaterialDocItem,
     decimal Quantity, string? Uom, string? Container, string? BillOfLading, string? InvoiceNumber,
     DateTime? DocumentDate, DateTime? PostingDate, decimal RemainingQty, string Source,
     DateTime CreatedAtUtc, string? CreatedByUsername, DateTime? ExpiryDate,
@@ -39,9 +39,9 @@ public sealed record CsvImportResult(int Imported, int Skipped);
 
 // ── Reversal-chain cancellation (pure, no DB — see ComputeReversalCancellations) ──
 
-public sealed record ReversalCancellationZeroedRow(long DeliveryId, string Material, string MaterialDocument, string MaterialDocItem, decimal Quantity);
+public sealed record ReversalCancellationZeroedRow(int DeliveryId, string Material, string MaterialDocument, string MaterialDocItem, decimal Quantity);
 
-public sealed record ReversalCancellationReviewRow(long DeliveryId, string Material, string MaterialDocument, string MaterialDocItem, decimal Quantity, decimal RemainingQty, string Reason);
+public sealed record ReversalCancellationReviewRow(int DeliveryId, string Material, string MaterialDocument, string MaterialDocItem, decimal Quantity, decimal RemainingQty, string Reason);
 
 public sealed record ReversalCancellationResult(IReadOnlyList<ReversalCancellationZeroedRow> Zeroed, IReadOnlyList<ReversalCancellationReviewRow> NeedsReview);
 
@@ -58,7 +58,7 @@ public sealed record VendorBalanceResult(
 // ── FEFO/FIFO/manual allocation proposal (pure, no DB — see BuildAllocationProposal) ──
 
 public sealed record AllocationProposalLine(
-    long DeliveryId, string Material, decimal QtyAllocated, string? InvoiceNumber,
+    int DeliveryId, string Material, decimal QtyAllocated, string? InvoiceNumber,
     DateTime? ExpiryDate, DateTime? DocumentDate, decimal RemainingBeforeAllocation);
 
 public sealed record AllocationProposalResult(string Method, IReadOnlyList<AllocationProposalLine> Lines, decimal UnallocatedQty, IReadOnlyList<ConsignmentDeliveryRow> OpenLines);
@@ -67,18 +67,18 @@ public sealed record ProposeDeclarationRequest(string Material, decimal QtyToDec
 
 // ── Declarations ───────────────────────────────────────────────────────────
 
-public sealed record CreateDeclarationLineRequest(long DeliveryId, string Material, decimal QtyAllocated);
+public sealed record CreateDeclarationLineRequest(int DeliveryId, string Material, decimal QtyAllocated);
 
 public sealed record CreateDeclarationRequest(string? AllocationMethod, List<CreateDeclarationLineRequest> Lines);
 
 public sealed record SetDeclarationLinesRequest(List<CreateDeclarationLineRequest> Lines);
 
 public sealed record ConsignmentDeclarationLineRow(
-    long DeclarationLineId, long DeliveryId, string Material, decimal QtyAllocated,
+    int DeclarationLineId, int DeliveryId, string Material, decimal QtyAllocated,
     string? InvoiceNumber, string? MaterialDocument, DateTime? DocumentDate, string? Uom, DateTime? ExpiryDate);
 
 public sealed record ConsignmentDeclarationSummaryRow(
-    long DeclarationId, long VendorId, string VendorName, string Status, string AllocationMethod, decimal TotalQty,
+    int DeclarationId, int VendorId, string VendorName, string Status, string AllocationMethod, decimal TotalQty,
     DateTime CreatedAtUtc, string? CreatedByUsername, DateTime? ConfirmedAtUtc, string? ConfirmedByUsername,
     string? SettlementDocumentNumber, decimal? SettlementReconciledQty, string? Notes);
 
@@ -88,13 +88,13 @@ public sealed record ConfirmDeclarationRequest(string? SettlementDocumentNumber,
 
 // ── Reassigning declarations off cancelled stock (pure algorithm + DB read/write — ported but not wired to any route, matching Node: real, unit-tested, but genuinely unrouted there too) ──
 
-public sealed record CancelledDeclarationLine(long DeclarationLineId, long DeclarationId, long CancelledDeliveryId, string Material, decimal QtyAllocated);
+public sealed record CancelledDeclarationLine(int DeclarationLineId, int DeclarationId, int CancelledDeliveryId, string Material, decimal QtyAllocated);
 
-public sealed record OpenDeliveryForReassignment(long DeliveryId, string Material, decimal RemainingQty, DateTime? ExpiryDate, DateTime? DocumentDate);
+public sealed record OpenDeliveryForReassignment(int DeliveryId, string Material, decimal RemainingQty, DateTime? ExpiryDate, DateTime? DocumentDate);
 
-public sealed record ReassignmentSplit(long DeliveryId, decimal Qty);
+public sealed record ReassignmentSplit(int DeliveryId, decimal Qty);
 
-public sealed record ReassignmentPlanItem(long DeclarationLineId, long DeclarationId, string Material, long CancelledDeliveryId, decimal TotalQty, IReadOnlyList<ReassignmentSplit> Splits, decimal Shortfall);
+public sealed record ReassignmentPlanItem(int DeclarationLineId, int DeclarationId, string Material, int CancelledDeliveryId, decimal TotalQty, IReadOnlyList<ReassignmentSplit> Splits, decimal Shortfall);
 
 public sealed record ReassignmentApplyResult(IReadOnlyList<ReassignmentPlanItem> Applied, IReadOnlyList<ReassignmentPlanItem> Skipped);
